6-size.c wot uasmntlong

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void)
 	long long int l_l_intType;
 	float fType;
 
-	printf("Size of char: %zu bytes\n", (unsigned long)sizeof(iType));
-	printf("Size of int: %zu bytes\n", (unsigned long)sizeof(cType));
-	printf("Size of long int: %zu bytes\n", (unsigned long)sizeof(l_intType));
-	printf("Size of long long int: %zu byte\n", (unsigned long)sizeof(l_l_intType));
-	printf("Size of float: %zu byte\n", (unsigned long)sizeof(fType));
+	printf("Size of char: %zu bytes\n", sizeof(iType));
+	printf("Size of int: %zu bytes\n", sizeof(cType));
+	printf("Size of long int: %zu bytes\n", sizeof(l_intType));
+	printf("Size of long long int: %zu byte\n", sizeof(l_l_intType));
+	printf("Size of float: %zu byte\n", sizeof(fType));
 	return (0);
 }
